requirements.txt
streamlit>=1.33.0
watchdog>=6.0.0
anthropic>=0.25.0
openai>=1.0.0
groq>=0.5.0
google-generativeai>=0.5.0
plotly>=5.0.0
